Move linters provider to standalone file
